more changes as per pylint recommendations
