+++ app/__init__.py
@@ -1,0 +1,1 @@
+'''A dummy file show that this folder is a module'''
