+++ app/database.py
@@ -6,9 +6,9 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy import create_engine
 
 postgres_host = os.environ.get("VACHAN_POSTGRES_HOST", "localhost")
-postgres_user = os.environ.get("VACHAN_POSTGRES_USER", "postgres") 
-postgres_database = os.environ.get("VACHAN_POSTGRES_DATABASE", "postgres") 
-postgres_password = os.environ.get("VACHAN_POSTGRES_PASSWORD", "secret") 
+postgres_user = os.environ.get("VACHAN_POSTGRES_USER", "postgres")
+postgres_database = os.environ.get("VACHAN_POSTGRES_DATABASE", "postgres")
+postgres_password = os.environ.get("VACHAN_POSTGRES_PASSWORD", "secret")
 
 SQLALCHEMY_DATABASE_URL = "postgresql://%s:%s@%s/%s"%(
 	postgres_user,postgres_password,postgres_host,postgres_database)
