+++ app/main.py
@@ -10,7 +10,7 @@ from fastapi.exceptions import RequestValidationError
 from starlette.exceptions import HTTPException as StarletteHTTPException
 from sqlalchemy.orm import Session
 
-import crud
+import crud 
 import db_models
 import schemas
 from database import SessionLocal, engine
